Cast args correctly to match format strings.

+++ sys/dev/pci/pccbb.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: pccbb.c,v 1.62 2001/04/12 18:18:31 thorpej Exp $	*/
+/*	$NetBSD: pccbb.c,v 1.63 2001/04/30 02:49:04 jmc Exp $	*/
 
 /*
  * Copyright (c) 1998, 1999 and 2000
@@ -463,7 +463,7 @@ pccbbattach(parent, self, aux)
 			    NULL)) {
 				printf("%s: can't map socket base address"
 				    " 0x%lx: io mode\n", sc->sc_dev.dv_xname,
-				    sockbase);
+				    (unsigned long)sockbase);
 				/* give up... allocate reg space via rbus. */
 				sc->sc_base_memh = 0;
 				pci_conf_write(pc, pa->pa_tag, PCI_SOCKBASE, 0);
@@ -3066,9 +3066,9 @@ pccbb_winlist_delete(head, bsh, size)
 		if ((chainp->wc_end - chainp->wc_start) != (size - 1)) {
 			printf("pccbb_winlist_delete: window 0x%lx size "
 			    "inconsistent: 0x%lx, 0x%lx\n",
-			    chainp->wc_start,
-			    chainp->wc_end - chainp->wc_start,
-			    size - 1);
+			    (unsigned long)chainp->wc_start,
+			    (unsigned long)(chainp->wc_end - chainp->wc_start),
+			    (unsigned long)(size - 1));
 			return 1;
 		}
